merapihkan file uts
